+++ src/components/molecules/LanguageSwitcher.tsx
@@ -1,22 +1,32 @@
 import { useTranslation } from 'react-i18next';
-import { ToggleButton, ToggleButtonGroup } from '@mui/material';
+import { Stack, ToggleButton, ToggleButtonGroup, Typography } from '@mui/material';
 import { MouseEvent } from 'react';
 import { LANGUAGES } from '../../constants/Languages';
 
 function LanguageSwitcher() {
-  const { i18n } = useTranslation();
-  const changeLanguage = (_: MouseEvent<HTMLElement>, lng: string) => {
-    void i18n.changeLanguage(lng);
-  };
+  const { t, i18n } = useTranslation();
+  function changeLanguage(_: MouseEvent<HTMLElement>, lng: string) {
+    if (lng !== null) void i18n.changeLanguage(lng);
+  }
 
   return (
-    <ToggleButtonGroup exclusive onChange={changeLanguage} value={i18n.language} size="small">
-      {Object.keys(LANGUAGES).map(languageKey => (
-        <ToggleButton key={languageKey} value={languageKey}>
-          {LANGUAGES[languageKey].nativeName}
-        </ToggleButton>
-      ))}
-    </ToggleButtonGroup>
+    <Stack direction="column">
+      <Typography variant="soloLabel" mb={1}>
+        {t('components.molecules.language_switcher.language')}
+      </Typography>
+      <ToggleButtonGroup
+        exclusive
+        value={i18n.language}
+        onChange={changeLanguage}
+        orientation="vertical"
+      >
+        {Object.keys(LANGUAGES).map(languageKey => (
+          <ToggleButton key={languageKey} value={languageKey}>
+            <Typography textTransform="capitalize">{LANGUAGES[languageKey].nativeName}</Typography>
+          </ToggleButton>
+        ))}
+      </ToggleButtonGroup>
+    </Stack>
   );
 }