+++ src/constants/ColorModes.ts
@@ -1,0 +1,2 @@
+export type IColorMode = 'light' | 'dark' | 'system';
+export const COLOR_MODES: IColorMode[] = ['light', 'dark', 'system'];