+++ src/contexts/MuiContextProvider.tsx
@@ -1,13 +1,17 @@
 import { createTheme, CssBaseline, ThemeProvider } from '@mui/material';
 import { ReactNode, useContext, useMemo } from 'react';
-import { MUI_THEME_OPTIONS } from '../constants/MuiThemeOptions';
+import { getMuiThemeOptions } from '../constants/MuiThemeOptions';
 import { ColorModeContext } from './ColorModeContextProvider';
 
 function MuiContextProvider({ children }: { children: ReactNode }) {
-  const { colorMode } = useContext(ColorModeContext);
+  const { muiColorMode } = useContext(ColorModeContext);
   const theme = useMemo(
-    () => createTheme({ ...MUI_THEME_OPTIONS, palette: { mode: colorMode } }),
-    [colorMode],
+    () =>
+      createTheme({
+        ...getMuiThemeOptions(muiColorMode === 'dark'),
+        palette: { mode: muiColorMode },
+      }),
+    [muiColorMode],
   );
 
   return (