+++ src/translations/es.ts
@@ -7,4 +7,22 @@ export default {
       go_home: 'Ir a inicio',
     },
   },
+  components: {
+    molecules: {
+      language_switcher: {
+        language: 'Idioma',
+      },
+      toggle_color_mode_button: {
+        theme: 'Tema',
+        light: 'Claro',
+        dark: 'Oscuro',
+        system: 'Sistema',
+      },
+    },
+    organisms: {
+      app_config_menu: {
+        title: 'Ajustes de la app',
+      },
+    },
+  },
 };